requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kafka-python==2.3.0
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
pandas==2.3.3
Pillow==12.1.0
py4j==0.10.9.9
pyspark==4.1.1
python-dateutil==2.9.0
pytz==2025.2
tzdata==2025.3
pyparsing==3.3.1
packaging==25.0
six==1.17.0
Flask==2.2.5
